Create MPU6050_HysteresisSmoothingPlotter.ino
Let's generate a short program for the Arduino Uno using the Adafruit_Sensor.h and Adafruit_MPU6050.h libraries which plots the output of the MPU6050 in the serial plotter and implements hysteresis smoothing. Suggest a file name upon completion and any additions which will enhance the functionality of the code.

Certainly! Hysteresis smoothing helps to reduce the noise in sensor readings. This is usually achieved by setting a threshold that must be exceeded for the sensor reading to be considered as changed.

Here's the code that reads values from MPU6050, implements hysteresis smoothing, and plots the output in the serial plotter.