Create MPU6050_DynamicHysteresisComparisonPlotter.ino
Let's refactor that code by adding a diagnostic function. Prompt the user to choose the hysteresis threshold and a set length of time to observe the effects. Prompt the user to set a length of time to observe the MPU sans hysteresis and have the code switch back and forth between these values. Suggest a filename upon completion. 

Certainly, here's the refactored code that prompts the user to set the hysteresis threshold, the length of time to observe the effects of hysteresis, and the length of time to observe the MPU without hysteresis. The program will switch back and forth between these two states.

This program will provide a better understanding of the effects of hysteresis on sensor readings by allowing the user to compare readings with and without hysteresis over user-defined time intervals.